fix: Created journey_planner namespace with itinerary result
