+++ app/graphql/types/tfl/entities/journey_planner/itinerary_result_type.rb
@@ -1,0 +1,17 @@
+module Types
+  module Tfl
+    module Entities
+      module JourneyPlanner
+        class ItineraryResultType < Types::BaseObject
+          field :stopMessages, [String, null: true], null: true
+          field :recommendedMaxAgeMinutes, Integer, null: true
+          field :lines, [Types::Tfl::Entities::LineType, null: true], null: true
+          field :journeys, [Types::JourneyType, null: true], null: true
+          field :journeyVector, Types::JourneyVectorType, null: true
+          field :searchCriteria, Types::SearchCriteriaType, null: true
+          field :cycleHireDockingStationData, Types::CycleHireDockingStationDataType, null: true
+        end
+      end
+    end
+  end
+end
