+++ spec/graphql/types/tfl/entities/journey_planner/itinerary_result_type_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe Types::ItineraryResultType do
+describe Types::Tfl::Entities::JourneyPlanner::ItineraryResultType do
   subject { described_class }
 
   it { is_expected.to have_field(:stopMessages).of_type('[String]') }
